@@ -1,10 +1,9 @@
 package ngsild
 
 import (
-	"log/slog"
 	"net/http"
 
-	"github.com/go-chi/chi/v5"
+	"github.com/diwise/service-chassis/pkg/infrastructure/o11y/logging"
 )
 
 // TODO: Load from file in file system instead of hardcoding a constant
@@ -14,17 +13,19 @@ const DefaultContext string = `{
     ]
 }`
 
-func NewServeContextHandler(logger *slog.Logger) http.HandlerFunc {
+func NewServeContextHandler() http.HandlerFunc {
 	responseBytes := []byte(DefaultContext)
 
 	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		contextID := chi.URLParam(r, "contextId")
+		ctx := r.Context()
+		contextID := r.PathValue("contextId")
 
 		if contextID != "default-context.jsonld" {
 			w.WriteHeader(http.StatusNotFound)
 			return
 		}
 
+		logger := logging.GetFromContext(ctx)
 		logger.Info("default context requested from client")
 
 		w.Header().Add("Content-Type", "application/json")
