@@ -3,7 +3,6 @@ package ngsild
 import (
 	"encoding/json"
 	"errors"
-	"log/slog"
 	"net/http"
 
 	"github.com/diwise/context-broker/internal/pkg/application/cim"
@@ -11,18 +10,15 @@ import (
 	ngsierrors "github.com/diwise/context-broker/pkg/ngsild/errors"
 	"github.com/diwise/context-broker/pkg/ngsild/types/entities"
 	"github.com/diwise/context-broker/pkg/ngsild/types/entities/decorators"
-	"github.com/diwise/service-chassis/pkg/infrastructure/o11y"
-	"github.com/diwise/service-chassis/pkg/infrastructure/o11y/tracing"
-	"go.opentelemetry.io/otel/attribute"
-	"go.opentelemetry.io/otel/trace"
+	"github.com/diwise/service-chassis/pkg/infrastructure/o11y/logging"
+	"go.opentelemetry.io/contrib/instrumentation/net/http/otelhttp"
 )
 
 // NewRetrieveAvailableEntityTypesHandler handles GET requests for the
 // entity types available in this NGSI-LD system
 func NewRetrieveAvailableEntityTypesHandler(
 	contextInformationManager cim.TypesRetriever,
-	authenticator auth.Enticator,
-	logger *slog.Logger) http.HandlerFunc {
+	authenticator auth.Enticator) http.HandlerFunc {
 
 	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
 		var err error
@@ -30,27 +26,25 @@ func NewRetrieveAvailableEntityTypesHandler(
 		ctx := r.Context()
 		tenant := GetTenantFromContext(ctx)
 
+		labeler, _ := otelhttp.LabelerFromContext(ctx)
+		defer func() { addLabelIfError(err, labeler) }()
+
 		propagatedHeaders := extractHeaders(r, "Accept", "Link")
-
-		ctx, span := tracer.Start(ctx, "retrieve-types",
-			trace.WithAttributes(attribute.String(TraceAttributeNGSILDTenant, tenant)),
-		)
-		defer func() { tracing.RecordAnyErrorAndEndSpan(err, span) }()
-
-		traceID, ctx, log := o11y.AddTraceIDToLoggerAndStoreInContext(span, logger, ctx)
 
 		detailsRequested := r.URL.Query().Get("details")
 		if detailsRequested != "" && detailsRequested != "false" {
 			err = errors.New("details not supported for /types")
-			ngsierrors.ReportNewBadRequestData(w, err.Error(), traceID)
+			ngsierrors.ReportNewBadRequestData(w, err.Error(), traceID(ctx))
 			return
 		}
+
+		log := logging.GetFromContext(ctx)
 
 		err = authenticator.CheckAccess(ctx, r, tenant, []string{})
 		if err != nil {
 			log.Warn("access not granted", "err", err.Error())
 			messageToSendToNonAuthenticatedClients := "not found"
-			ngsierrors.ReportNotFoundError(w, messageToSendToNonAuthenticatedClients, traceID)
+			ngsierrors.ReportNotFoundError(w, messageToSendToNonAuthenticatedClients, traceID(ctx))
 			return
 		}
 
@@ -62,7 +56,7 @@ func NewRetrieveAvailableEntityTypesHandler(
 		availableTypes, err := contextInformationManager.RetrieveTypes(ctx, tenant, propagatedHeaders)
 		if err != nil {
 			log.Error("query entities failed", "err", err.Error())
-			mapCIMToNGSILDError(w, err, traceID)
+			mapCIMToNGSILDError(w, err, traceID(ctx))
 			return
 		}
 
@@ -76,7 +70,7 @@ func NewRetrieveAvailableEntityTypesHandler(
 
 		if err != nil {
 			log.Error("retrieve entity types: failed to marshal type list to json", "err", err.Error())
-			mapCIMToNGSILDError(w, err, traceID)
+			mapCIMToNGSILDError(w, err, traceID(ctx))
 			return
 		}
 
